requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
fastapi==0.72.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
pydantic==1.9.0
PyMySQL==1.0.2
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.34
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
